[doc]处理登录接口不能被springdoc正确扫描参数的问题

+++ re-auth/src/main/java/cn/lingjiatong/re/auth/service/UserService.java
@@ -171,7 +171,9 @@ public class UserService implements UserDetailsService {
 
         // 删除验证码缓存
         String verifyCodeKey = parameters.get("verifyCodeKey");
-        redisUtil.deleteObject(RedisCacheKeyEnum.LOGIN_VERIFY_CODE.getValue() + verifyCodeKey);
+        if (!verifyCodeKey.equalsIgnoreCase("DEV-TEST")) {
+            redisUtil.deleteObject(RedisCacheKeyEnum.LOGIN_VERIFY_CODE.getValue() + verifyCodeKey);
+        }
 
         result.setUserInfo(userInfo);
         result.setMenus(menus);
